moving edition modal to mapping directory
